Clean up code related to loading keyrings
The operations to load a public and a private keyring are identical, so
that part can be moved to a common function.

Some operations that need to load keyrings have unnecessary or redundant
paths of code. Remove that.

Adding keys is opening the keyring, keeping it open, and then opening it
again. Remove the redundant open.

When importing a file with a key, try to read binary data from it, and
if that fails retry reading ascii-armored data.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>